refactor: move JobCategory enum to common domain
JobCategory is now shared between profile and interview domains,
so moved to common.domain to avoid cross-domain imports.
