+++ src/main/java/wlsh/project/intervai/profile/domain/CreateProfileCommand.java
@@ -1,6 +1,7 @@
 package wlsh.project.intervai.profile.domain;
 
 import java.util.List;
+import wlsh.project.intervai.common.domain.JobCategory;
 
 public record CreateProfileCommand(
         JobCategory jobCategory,
