+++ src/test/java/wlsh/project/intervai/profile/presentation/ProfileControllerTest.java
@@ -13,7 +13,7 @@ import wlsh.project.intervai.common.exception.CustomException;
 import wlsh.project.intervai.common.exception.ErrorCode;
 import wlsh.project.intervai.profile.application.ProfileService;
 import wlsh.project.intervai.profile.domain.CareerLevel;
-import wlsh.project.intervai.profile.domain.JobCategory;
+import wlsh.project.intervai.common.domain.JobCategory;
 import wlsh.project.intervai.profile.domain.Profile;
 import wlsh.project.intervai.profile.domain.UpdateProfileCommand;
 import wlsh.project.intervai.profile.presentation.dto.UpdateProfileRequest;
